Revert "Merge remote-tracking branch 'upstream/master'"
This reverts commit 9af10c7529371e0da09fb8522ade082a913b774c, reversing
changes made to 72bd081e1b4a189aa454d79db3e42f3015f7a91f.

+++ gframe/CGUICustomText/CGUICustomText.h
@@ -182,14 +182,14 @@ private:
 
 	// auto scrolling related functions and variables
 	CTEXT_SCROLLING_TYPE scrolling;
-	float maxFrame;
+	int maxFrame;
 	float curFrame;
 	float frameTimer;
-	float forcedSteps;
+	int forcedSteps;
 	float forcedStepsRatio;
 	float animationStep;
-	float animationWaitStart;
-	float animationWaitEnd;
+	int animationWaitStart;
+	int animationWaitEnd;
 	bool increasingFrame;
 	bool waitingEndFrame;
